Factor out daml2ts utilities
changelog_begin
changelog_end
