+++ daml-assistant/integration-tests/BUILD.bazel
@@ -123,6 +123,7 @@ da_haskell_test(
         "//:sdk-version-hs-lib",
         "//compiler/damlc/daml-opts:daml-opts-types",
         "//daml-assistant/daml-helper:daml-helper-lib",
+        "//language-support/ts/codegen:daml2ts-helpers",
         "//libs-haskell/bazel-runfiles",
         "//libs-haskell/da-hs-base",
     ],
